REST/requirements.txt
django
python-dotenv
djangorestframework
djangorestframework-simplejwt
drf-spectacular